Refactor: passport 성격별로 loader 분리
